+++ packages/example/dev/main_dev.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_wright_example/app.dart';
 import 'package:flutter_wright_example/router/app_router.dart';
@@ -19,6 +20,7 @@ import 'package:flutter_wright_sdk/flutter_wright_sdk.dart';
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await FlutterWright.start(
+    enabled: kDebugMode,
     navigatorKey: FlutterWright.navigatorKey,
     routes: AppRouter.names,
   );